Extract foundation core snapshot mapper
