+++ tests/unit/test_foundation_service.py
@@ -729,83 +729,6 @@ async def test_foundation_workspace_rejects_portfolio_identity_upstream_error():
     assert "secret-token" not in exc_info.value.detail
 
 
-def test_foundation_parses_defensive_snapshot_branches():
-    service = FoundationService(
-        lotus_core_query_client=_StubLotusCoreQueryClient({"items": []}, {}),
-        analytics_client=_StubAnalyticsClient(200, {}),
-        dpm_client=_StubDpmClient(200, {}),
-        reporting_client=_StubReportingClient(200, {}),
-    )
-
-    portfolio, summary, allocations, top_positions, as_of_date = service._parse_core_snapshot(
-        fallback_portfolio_id="PF_FALLBACK",
-        fallback_as_of_date="2026-03-31",
-        portfolio_payload={"name": "Fallback Name"},
-        payload={
-            "portfolio_id": "PF_FALLBACK",
-            "as_of_date": "2026-03-30",
-            "valuation_context": {"portfolio_currency": "EUR"},
-            "sections": {
-                "positions_baseline": [
-                    "skip-me",
-                    {"security_id": None, "asset_class": "Alternatives", "value_base": "125.55"},
-                    {"security_id": "EQ_1", "valuation": {"market_value_base": "374.45"}},
-                ],
-                "portfolio_totals": [],
-                "instrument_enrichment": [
-                    "skip-me",
-                    {"security_id": "EQ_1", "asset_class_name": "Equity"},
-                ],
-            },
-        },
-    )
-
-    assert portfolio.portfolio_id == "PF_FALLBACK"
-    assert portfolio.display_name == "Fallback Name"
-    assert portfolio.base_currency == "EUR"
-    assert summary.market_value_base == 0.0
-    assert len(top_positions) == 2
-    assert summary.position_count == 2
-    assert [bucket.asset_class for bucket in allocations] == ["Alternatives", "Equity"]
-    assert allocations[0].weight_pct is None
-    assert allocations[1].market_value_base == 374.45
-    assert as_of_date == "2026-03-30"
-
-
-def test_foundation_snapshot_parser_ignores_legacy_nested_portfolio_and_metadata():
-    service = FoundationService(
-        lotus_core_query_client=_StubLotusCoreQueryClient({"items": []}, {}),
-        analytics_client=_StubAnalyticsClient(200, {}),
-        dpm_client=_StubDpmClient(200, {}),
-        reporting_client=_StubReportingClient(200, {}),
-    )
-
-    portfolio, _summary, _allocations, _top_positions, as_of_date = service._parse_core_snapshot(
-        fallback_portfolio_id="PF_IGNORED",
-        fallback_as_of_date="2026-03-31",
-        portfolio_payload={
-            "portfolio_id": "PF_IDENTITY",
-            "portfolio_name": "Identity Name",
-            "base_currency": "USD",
-        },
-        payload={
-            "portfolio_id": "PF_SNAPSHOT",
-            "as_of_date": "2026-03-30",
-            "portfolio": {"portfolio_id": "PF_LEGACY", "portfolio_name": "Legacy Name"},
-            "metadata": {"business_date": "1999-01-01"},
-            "sections": {
-                "positions_baseline": [],
-                "portfolio_totals": {},
-                "instrument_enrichment": [],
-            },
-        },
-    )
-
-    assert portfolio.portfolio_id == "PF_SNAPSHOT"
-    assert portfolio.display_name == "Identity Name"
-    assert as_of_date == "2026-03-30"
-
-
 def test_foundation_parses_optional_result_edge_cases():
     service = FoundationService(
         lotus_core_query_client=_StubLotusCoreQueryClient({"items": []}, {}),
@@ -875,25 +798,13 @@ def test_foundation_parses_optional_result_edge_cases():
     assert partial_failures[0].error_code == "HTTP_503"
 
 
-def test_foundation_extracts_market_value_and_workflow_cues():
+def test_foundation_builds_workflow_cues():
     service = FoundationService(
         lotus_core_query_client=_StubLotusCoreQueryClient({"items": []}, {}),
         analytics_client=_StubAnalyticsClient(200, {}),
         dpm_client=_StubDpmClient(200, {}),
         reporting_client=_StubReportingClient(200, {}),
     )
-
-    assert service._extract_market_value({"valuation": {"market_value": "88.10"}}) == 88.1
-    assert (
-        service._extract_market_value(
-            {"valuation": {"market_value_base": "bad"}, "current_value": "45.25"}
-        )
-        == 45.25
-    )
-    assert (
-        service._extract_market_value({"market_value_base": "bad", "value_base": "17.75"}) == 17.75
-    )
-    assert service._extract_market_value({"valuation": {"market_value_base": "bad"}}) is None
 
     cues = service._build_workflow_cues("PF_1001")
     assert [cue.key for cue in cues] == ["performance", "risk", "proposal"]
